requirements.txt
google-generativeai==0.2.1
scikit-learn==1.0.2
sentence-transformers==2.2.0
torch==1.10.0